XUSG update for making uint8_t counters for FrameCount.

+++ IrradianceMap/IrradianceMap.h
@@ -44,7 +44,7 @@ public:
 	virtual void ParseCommandLineArgs(wchar_t* argv[], int argc);
 
 private:
-	static const uint32_t FrameCount = 3;
+	static const uint8_t FrameCount = 3;
 
 	std::shared_ptr<XUSG::DescriptorTableCache> m_descriptorTableCache;
